Refactor CSS padding in main and post pages
This commit refactors the CSS padding in the main and post pages. In the main page, the padding of the .main class is updated to include both left and right padding of 40px. In the post page, the padding of the .post-info class is updated to include a top padding of 50px. These changes improve the spacing and layout of the pages.